+++ selendroid-client/src/main/java/io/selendroid/DriverCommand.java
@@ -1,0 +1,10 @@
+package io.selendroid;
+
+public enum DriverCommand {
+  SEND_KEYS_TO_ELEMENT(org.openqa.selenium.remote.DriverCommand.SEND_KEYS_TO_ELEMENT);
+  DriverCommand(String command) {
+    this.command = command;
+  }
+
+  public String command;
+}