+++ selendroid-server/src/main/java/io/selendroid/server/model/Session.java
@@ -13,17 +13,31 @@
  */
 package io.selendroid.server.model;
 
+import java.util.HashMap;
+import java.util.Map;
+
+import org.json.JSONException;
 import org.json.JSONObject;
 
 public class Session {
+  public static final String SEND_KEYS_TO_ELEMENT = "sendKeysToElement";
+  public static final String NATIVE_EVENTS_PROPERTY = "nativeEvents";
+
   private JSONObject capabilities = null;
   private KnownElements knownElements = null;
   private String sessionId = null;
+  private Map<String, JSONObject> commandConfiguration;
 
   public Session(JSONObject capabilities, String sessionId) {
     this.capabilities = capabilities;
     this.sessionId = sessionId;
     this.knownElements = new KnownElements();
+    this.commandConfiguration = new HashMap<String, JSONObject>();
+    JSONObject configJsonObject = new JSONObject();
+    try {
+      configJsonObject.put(NATIVE_EVENTS_PROPERTY, true);
+    } catch (JSONException e) {}
+    this.commandConfiguration.put(SEND_KEYS_TO_ELEMENT, configJsonObject);
   }
 
   public JSONObject getCapabilities() {
@@ -37,4 +51,20 @@ public class Session {
   public String getSessionId() {
     return sessionId;
   }
+
+  public void setCommandConfiguration(String command, JSONObject config) {
+    if (commandConfiguration.containsKey(command)) {
+      commandConfiguration.remove(command);
+    }
+    commandConfiguration.put(command, config);
+  }
+
+  public JSONObject getCommandConfiguration(String command) {
+    if (commandConfiguration.containsKey(command)) {
+      return commandConfiguration.get(command);
+    }
+    return null;
+  }
+
+
 }