+++ changelog.md
@@ -2,6 +2,7 @@
 ---
 - Rotation Support
 - Fixes: [#268](../../issues/268), [#284](../../issues/284)
+- Adding support for disabling native events for sendKeys command to support e.g. German umlauts (#7,#110)
 
 0.8.0
 ---